update text client for computer player and some refactoring
