+++ lib/game/easy_strategy.rb
@@ -1,0 +1,14 @@
+module Game
+  class EasyStrategy
+    attr_reader :name, :type
+
+    def initialize()
+      @name = :easy
+      @type = :computer
+    end
+
+    def compute_move(board, position = nil)
+      board.available_positions.sample
+    end
+  end
+end
