+++ lib/game/human_strategy.rb
@@ -1,0 +1,14 @@
+module Game
+  class HumanStrategy
+    attr_reader :name, :type
+
+    def initialize()
+      @name = :human
+      @type = :user
+    end
+
+    def compute_move(board, position)
+      position
+    end
+  end
+end
