--- lib/human_strategy.rb
@@ -1,13 +1,0 @@
-class HumanStrategy
-  attr_reader :name, :type
-
-  def initialize()
-    @name = :human
-    @type = :user
-  end
-
-  def compute_move(board, position)
-    position
-  end
-end
-
